requirements.txt
numpy
asap3
ase
torch
tqdm
pymatgen
lz4
ipykernel
torch>=1.9.0
numpy>=1.20.0
ase>=3.22.0
pymatgen>=2022.0.0
omegaconf>=2.1.0
hydra-core>=1.1.0
pandas>=1.3.0
scipy>=1.7.0
